requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.4
distro==1.9.0
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.6.1
openai==1.52.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
python-magic==0.4.27
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
